+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fix tests to work with newest changes.
 - Fix cached null socket file name after fresh install bug. 
   - Now on game server start, will purge cache entry if socket file name is null.